refactor: move `fromThemeUI` to its own file
