+++ packages/glaze/src/index.ts
@@ -1,9 +1,10 @@
+export { fromThemeUI } from './converters/fromThemeUI';
 export { setLogger } from './logger';
 export { defaultTokens } from './presets/defaultTokens';
 export { modularScale } from './scales';
 export { VirtualStyleInjector } from './StyleInjector';
 export { StyleInjectorProvider } from './StyleInjectorContext';
-export { createTheme, fromThemeUI } from './theme';
+export { createTheme } from './theme';
 export { ThemeProvider, useTheme } from './ThemeContext';
 export type { ThemedStyle } from './useStyling';
 export { useStyling } from './useStyling';
